+++ Assets/ArtNet/Runtime/Scripts/Core/Enums/OpCode.cs
@@ -5,6 +5,12 @@ namespace ArtNet.Enums
         Poll = 0x2000,
         PollReply = 0x2100,
         Dmx = 0x5000,
-        Sync = 0x5200
+        Sync = 0x5200,
+        Address = 0x6000,
+        TodRequest = 0x8000,
+        TodData = 0x8100,
+        TodControl = 0x8200,
+        Rdm = 0x8300,
+        TimeCode = 0x9700
     }
 }